Изменил модель Event, добавив лист EventMember Добавил Сревис расчета распределения трат DivisionOfExpenseService добавил в скрипт миграции сущность EventMember, джобавил дефолтные значения даты и суммы, убрал Double заменив на numeric(19.2)

+++ src/main/java/ru/division/of/expenses/app/models/PartitialPayer.java
@@ -6,6 +6,7 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -16,7 +17,7 @@ public class PartitialPayer extends AbstractEntity {
     private User user;
 
     @Column
-    private Double coefficient;
+    private BigDecimal coefficient;
 
     @ManyToOne
     @JoinColumn(name = "expense_id")
